FIX: ODY-738, Long is now default for numbers without dot (instead of int)
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@20126 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: c4edfb3e779d998b8fb74fc69b3bd503bf6a3b1b

+++ common/mep/src/de/uniol/inf/is/odysseus/mep/impl/MEPImpl.java
@@ -445,7 +445,7 @@ public class MEPImpl/*@bgen(jjtree)*/implements MEPImplTreeConstants, MEPImplCon
         t = jj_consume_token(INTEGER);
     jjtree.closeNodeScope(jjtn000, true);
     jjtc000 = false;
-    jjtn000.setValue(Integer.parseInt(t.image));
+    jjtn000.setValue(Long.parseLong(t.image));
         break;
       case FLOAT:
         t = jj_consume_token(FLOAT);
